Refactor work generator
