+++ mirrulations-work-generator/.coveragerc
@@ -1,0 +1,6 @@
+[report]
+omit =
+    *wsgi.py
+exclude_lines =
+    pragma: no cover
+    if __name__ == .__main__.:
