+++ mirrulations-work-generator/tests/test_job_queue.py
@@ -1,6 +1,6 @@
 
 from fakeredis import FakeRedis
-from c21server.work_gen.job_queue import JobQueue
+from mirrgen.job_queue import JobQueue
 
 
 def test_first_job_added_with_id_0():
