+++ mirrulations-work-generator/tests/test_regulations_api.py
@@ -1,7 +1,6 @@
 import pytest
 from requests import HTTPError
-from c21server.work_gen.regulations_api import RegulationsAPI, \
-    MIN_DELAY_BETWEEN_CALLS
+from mirrgen.regulations_api import RegulationsAPI, MIN_DELAY_BETWEEN_CALLS
 
 
 def test_download_returns_json(requests_mock, mocker):
