+++ mirrulations-work-generator/tests/test_workgen.py
@@ -1,9 +1,9 @@
 from fakeredis import FakeRedis
-from c21server.work_gen.job_queue import JobQueue
-from c21server.work_gen.work_generator import WorkGenerator
-from c21server.work_gen.regulations_api import RegulationsAPI
-from c21server.work_gen.mock_dataset import MockDataSet
-from c21server.work_gen.mock_data_storage import MockDataStorage
+from mirrgen.job_queue import JobQueue
+from mirrgen.work_generator import WorkGenerator
+from mirrgen.regulations_api import RegulationsAPI
+from mirrgen.mock_dataset import MockDataSet
+from mirrgen.mock_data_storage import MockDataStorage
 
 
 def test_work_generator_single_page(requests_mock, mocker):
